+++ rana_qgis_plugin/utils/api.py
@@ -656,14 +656,20 @@ def get_threedi_schematisation(descriptor_id: str):
     raise RanaFetchError(f"Failed to retrieve schematisation: {error}", url, {})
 
 
-def copy_threedi_schematisation(project_id: str, schematisation_id: str, path: str):
+def copy_threedi_schematisation(
+    project_id: str, schematisation_id: str, revision_id: int, path: str
+):
     authcfg_id = get_authcfg_id()
     tenant = get_tenant_id()
     url = (
         f"{api_url()}/tenants/{tenant}/projects/{project_id}/model-schematisations/copy"
     )
     network_manager = NetworkManager(url, authcfg_id)
-    params = {"schematisation_id": schematisation_id, "path": path}
+    params = {
+        "schematisation_id": schematisation_id,
+        "revision_id": revision_id,
+        "path": path,
+    }
     status, error = network_manager.post(params)
     if status:
         return network_manager.content